feat(all): initial commit for managedSchema
*Motivation*

Managed schema is a `recordSpace` where the schema is directly handled
by the record-store. This allow to offer a dedicated grpc service that
you can use to shim something, or offer another type of dev experience.

The first Managed schema is the KV, which offers a Key/Value experience.
It has his own grpc service that will be easier to implement. Others may
follow.